fix: 메시지 전송 로직 수정
- i/o cost가 큰 작업이므로 비동기적으로 수행
