+++ src/main/java/com/softeer/podo/verification/controller/VerificationController.java
@@ -1,6 +1,7 @@
 package com.softeer.podo.verification.controller;
 
 import com.softeer.podo.common.response.CommonResponse;
+import com.softeer.podo.verification.exception.MessageSendFailException;
 import com.softeer.podo.verification.facade.VerificationFacade;
 import com.softeer.podo.verification.model.dto.request.CheckVerificationRequestDto;
 import com.softeer.podo.verification.model.dto.response.CheckVerificationResponseDto;
@@ -10,6 +11,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
+import java.util.concurrent.CompletableFuture;
 
 @RestController
 @RequestMapping("/verification")
@@ -20,9 +22,12 @@ public class VerificationController {
 
     @PostMapping("/claim")
     @Operation(summary = "전화번호로 인증번호 요청")
-    public CommonResponse<String> claimVerificationCode(@Valid @RequestBody ClaimVerificationRequestDto dto) {
-        verificationFacade.claimVerificationCode(dto);
-        return new CommonResponse<>("인증번호 발송 완료");
+    public CompletableFuture<CommonResponse<Void>> claimVerificationCode(@Valid @RequestBody ClaimVerificationRequestDto dto) {
+        return verificationFacade.claimVerificationCode(dto)
+                .thenApply(result -> new CommonResponse<>(result))
+                .exceptionally(ex -> {
+                    throw new MessageSendFailException("[비동기 에러] 메시지 발송에 실패했습니다");
+                });
     }
 
     @PostMapping("/check")
